requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
ddgs==9.6.0
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.11.2
django-cors-headers==4.9.0
django-filter==25.2
django-postgres-copy==2.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
duckduckgo_search==8.1.1
fastuuid==0.13.5
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.77.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.4
openai==2.0.1
packaging==25.0
pillow==11.3.0
primp==0.15.0
propcache==0.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
smolagents==1.22.0
sniffio==1.3.1
socksio==1.0.0
sqlparse==0.5.3
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Wikipedia-API==0.8.1
yarl==1.20.1
zipp==3.23.0
zope.interface==8.0.1
